+++ CHANGELOG.md
@@ -11,6 +11,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Allow passing in a mermaid `configFile` as an object to `mermaid-cli`.
 
+### Fixed
+
+- Fix wide SVGs (e.g. gitgraphs) being cut at 300px.
+  Automatically replaces SVG `width=100%` to the width in pixels.
+  Most browsers will cut SVGs to 300px. (closes [#7](https://github.com/aloisklink/remark-mermaid-dataurl/issues/7))
+
 ## [1.0.2] - 2021-03-29
 
 ### Fixed